requirements.txt
python-telegram-bot>=5.0
requests>=2.0
dpath>=1.3.0